fastapi/requirements.txt
# routers/requirements.txt

# Основные зависимости API
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.22.0,<0.33.0
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.28.0,<0.30.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
temporalio>=1.0.0,<2.0.0
redis>=4.5.0,<5.0.0

# Зависимости для тестирования
pytest>=8.0
pytest-asyncio>=0.20
httpx>=0.28.0,<0.29.0
starlette>=0.40.0,<0.41.0

#Исправление для Pydantic V2
pydantic[v1]>=2.0.0,<3.0.0
email-validator>=2.0.0,<3.0.0